+++ src/components/auth/index.ts
@@ -1,0 +1,4 @@
+export { default as AuthContents } from './AuthContents';
+export { default as LogoAndMessage } from './LogoAndMessage';
+export { default as InterchangeButton } from './InterchangeButton';
+export { default as AuthForm } from './AuthForm';
