+++ src/schemas/index.ts
@@ -1,0 +1,2 @@
+export { default as loginSchema } from './loginSchema';
+export { default as signupSchema } from './signupSchema';
